Move StringTag component to src/components/elements/reactions/propertiesTab/
