+++ app/packs/src/components/elements/reactions/propertiesTab/ReactionDetailsProperties.js
@@ -9,7 +9,7 @@ import 'moment-precise-range-plugin';
 import Clipboard from 'clipboard';
 import { dangerousProductsOptions } from 'src/components/staticDropdownOptions/options';
 import ReactionDetailsMainProperties from 'src/components/elements/reactions/ReactionDetailsMainProperties';
-import StringTag from 'src/components/StringTag';
+import StringTag from 'src/components/elements/reactions/propertiesTab/StringTag';
 import { solventsTL } from 'src/utilities/reactionPredefined';
 import OlsTreeSelect from 'src/components/OlsComponent';
 import { permitOn } from 'src/components/common/uis';
